requirements.txt
# Core Dependencies
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
requests==2.31.0

# AI/ML Stack
langchain-core
langchain-google-genai
google-generativeai


typing-extensions==4.11.0
python-multipart==0.0.9